modularização das classes
